+++ packages/conc/lib/Polysemy/Conc/Effect/Semaphore.hs
@@ -1,0 +1,7 @@
+module Polysemy.Conc.Effect.Semaphore where
+
+data Semaphore :: Effect where
+  Wait :: Semaphore m ()
+  Signal :: Semaphore m ()
+
+makeSem ''Semaphore