+++ packages/conc/lib/Polysemy/Conc/Semaphore.hs
@@ -1,0 +1,10 @@
+module Polysemy.Conc.Semaphore (
+  Semaphore,
+  interpretSemaphoreQ,
+  interpretSemaphoreT,
+  signal,
+  wait,
+) where
+
+import Polysemy.Conc.Effect.Semaphore (Semaphore, signal, wait)
+import Polysemy.Conc.Interpreter.Semaphore (interpretSemaphoreQ, interpretSemaphoreT)