+++ packages/conc/changelog.md
@@ -3,6 +3,7 @@
 * Add `PScoped`, a variant of `Scoped` that allows passing arguments to the scope.
 * Add `Queue.tryReadMaybe`, a variant of `readMaybe` that does not wait.
 * Add variants of the combined `Scoped` interpreters in which the `Resumable` is inside the scope.
+* Add `Semaphore`, abstracting `QSem` or `TSem`.
 
 # 0.9.0.0